Refactor elements with embedded HTML
