+++ src/Facebook/InstantArticles/Elements/Interactive.php
@@ -21,7 +21,7 @@ use Facebook\InstantArticles\Validators\Type;
  *
  * @see {link:https://developers.intern.facebook.com/docs/instant-articles/reference/interactive}
  */
-class Interactive extends Element implements Container
+class Interactive extends ElementWithHTML implements Container
 {
     const NO_MARGIN = 'no-margin';
     const COLUMN_WIDTH = 'column-width';
@@ -47,11 +47,6 @@ class Interactive extends Element implements Container
      * @see Interactive::COLUMN_WIDTH
      */
     private $width;
-
-    /**
-     * @var \DOMNode The HTML of the content.
-     */
-    private $html;
 
     private function __construct()
     {
@@ -132,21 +127,6 @@ class Interactive extends Element implements Container
     }
 
     /**
-     * Sets the unescaped HTML of your interactive graphic.
-     *
-     * @param \DOMNode $html The unescaped HTML of your interactive graphic.
-     *
-     * @return $this
-     */
-    public function withHTML($html)
-    {
-        Type::enforce($html, 'DOMNode');
-        $this->html = $html;
-
-        return $this;
-    }
-
-    /**
      * @return Caption the caption element
      */
     public function getCaption()
@@ -176,14 +156,6 @@ class Interactive extends Element implements Container
     public function getWidth()
     {
         return $this->width;
-    }
-
-    /**
-     * @return \DOMNode unescaped html
-     */
-    public function getHtml()
-    {
-        return $this->html;
     }
 
     /**
